+++ packages/runtime/src/index.ts
@@ -1,5 +1,5 @@
 import "@/styles/index.less";
-import { semverToIdentifier, getMajorVersion } from "widget-up-utils";
+import { convertSemverVersionToIdentify, getMajorVersion } from "widget-up-utils";
 
 export { globalEventBus } from "./events";
 export { start } from "./start";
@@ -8,6 +8,6 @@ export { createWindow } from "widget-up-utils";
 export type * from "@/types";
 
 export const utils = {
-  semverToIdentifier,
+  convertSemverVersionToIdentify,
   getMajorVersion,
 };